Remove pyparsing: not going to work.
Pyparsing is the wrong tool for the job, or I'm not able to do it right.
There are problems when mixing status definitions with other definitions
in between, it swallows things without raising errors, it is not really
line based etc.

Parsing the workflow definitions with an own parser / grammer / lexer is
overkill - using config parser should get us far enough..