Fix casing in role translation proposals.
The problem is that we only have the case insensitive customer role after parsing,
but we need to display the case sensitive original text as translation proposal.

Since lawgiver heavily relies on the role beeing case insensitive and it is used in
dicts as keys, I added a little hack which uses the lowercase string but keeps the
original for display in the view.