Fix action group translations for default action groups.
When a package, which extends default action groups, is loaded before
lawgiver, it might define the action group in translation domain other
than "ftw.lawgiver", resulting in an untranslated action group.

This adds a fallback to the "ftw.lawgiver" domain if the action group
is not translated in its default domain.