German parsing: fix parsing of action group "Inhalte konfigurieren".
The problem is that "Inhalte" is used in general statements, where it
is not included in the action group name, e.g:

  "Ein Administrator kann Inhalte immer bearbeiten."
  should parse to: ("Administrator", "bearbeiten")

but we also have an action groups which include it the name:

  "Der Administrator kann Inhalte konfigurieren."
  which should parse to: ("Administrator", "Inhalte konfigurieren")

We cannot remove "Inhalte" in this case because "konfigurieren" has a
different, incorrect meaning alone.

For fixing this we use a negative lookahead which does not remove
"Inhalte" when it is followed by " konfigurieren".