Add support for assigning permissions to multiple action groups.
There are situations when a permission must be mapped to multiple action groups.
For example when having multiple "add" action groups (for adding different types),
the "Add portal content" permission should be in each such action groups.

This can now be achieved by using the optional "move" flag (default: True) of the
"map_permissions" directive.
When using move="False", the permission is not removed from the original action group
but only added to the new one.