Drop Plone 4.3.4 specific tests.
The Plone 4.3.4 specific tests were added because plone.app.iterate was
added as a dependency in Plone 4.3.5, which introduced new permissions.

Since we now have the `Plone` egg as dependency, we always have those
dependencies installed and therefore have those permissions.
This makes the definition.xml of the example-workflow under test be the
same again for Plone 4.3.4 and Plone 4.3.5, making the Plone 4.3.4 test
obsolete.