Make collective.deletepermission dependency optional.
The ``collective.deletepermission`` dependency is optional from now on.
There is a new ``ftw.lawgiver [deletepermission]`` extras.

``collective.deletepermission`` is still considered a good thing, but it
is quite invasive as it patches various parts of Plone and introduces a
new permission. It may not be helpful in all projects.

The goal is to make lawgiver useable in an environment where
``collective.deletepermission`` cannot be installed.

The "delete objects" permission is mapped differently depending on
whether ``collective.deletepermission`` is installed, since the meaning
of the permission changes.