Fix review state translation in sharing page role description popup.
With ``ftw.lawgiver==1.6.1`` the translation of the review state was
changed from translating the state id to translating the state title.
This has changed because Plone also uses the review state title.

The role description popup of the sharing page was not updated in that
change and thus did not correctly translate the review state title in
the table heading.

This change changes the translation behavior of the state in the popup
so that we try to translate it the correct way (using the state title)
and gracefully fallback to the old behavior by using it as translation
default.