Also translate states and transitions with their IDs.
In version 1.6.1 we have switched from title based translations to
title based, for states and transitions.

The reason for switching to title based transitions was that some UI
components of Plone do translate the title string in the "plone"
translation domain.
But in catalog based listing usually the id of the statis is
translated in the "plone" domain.

For supporting as many places as possible we are now translating
states and transition by their titles as well as by their IDs.