Testing: improve diff output in workflow test for unittest
We have switched from unittest2 to unittest recently.
Both have a "assertMultiLineEqual", but the latter does handle large
datasets very well: unittest.assertMultiLineEqual just displays all on
one line when a definition.xml is diffed; what we do.

Therefore the XML diffing is reimplemented so that a unified diff is
printed in multiple lines so that the assertion message is readable
again for a human beeing.